requirements.txt
discord.py
searchtweets